Mended a strange bug freezing calculation on Stallo
This bug would cause the calculation to freeze at strange stages
when using an Intel compilation setup on Stallo,
and is somehow related to how and when perturbations are put in
standard order. The offending part was reverted to previous state,
and this made the bug go away although its cause is not understood.